refactor: move student management component
